Add files via upload
enrollment.py connects to the db, the user id is hardcoded & set to me. it reads the professors.txt file and matches all the professors from that list to the professor id's from the professors table in the database. and then it assigns my id to those professor ids and they are uploaded into the user_professors table